backend/requirements.txt
flask
flask-smorest
flask-cors
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
passlib
flask-migrate
gunicorn
requests
psycopg2-binary
redis
jinja2